Fixed finetune.py
I do not want to use a pip env. Especially when you tell me to compile
deps from source from git except you specified that like a year ago so
now the version does not match what you were using but you did not write
down the version number. So now I had to update your stupid script so
that it will run with whatever the latest is. Do not ask me what version
I am using at this moment because I do not know. But it runs now. I had
to completely remove the wandb dep because it fails to import. I think
because of a version mismatch with protobuf but it is unclear. I will
add version tags later.

I use Arch btw.